+++ src/App.jsx
@@ -4,6 +4,7 @@ import Board from "./components/Board";
 function App() {
   return (
     <div>
+        <h4>Tic-tac-toe</h4>
         <Board />
     </div>
   );
