+++ src/constants.jsx
@@ -1,4 +1,4 @@
-export const BOARD_SIZE = 4;
+export const BOARD_SIZE = 3;
 
 export const HORIZONTALLY = 'HORIZONTALLY';
 export const VERTICALLY = 'VERTICALLY';
