+++ src/forms/BoardSizeForm.jsx
@@ -1,45 +1,20 @@
-import React, {useState} from 'react';
-import * as yup from 'yup';
+import React from 'react';
+import {BOARD_SIZE} from "../constants";
 
 const BoardSizeForm = ({boardSize, setBoardSize, handleChangeBoardSize, winner}) => {
-    const [error, setError] = useState(false);
-    const [errorMessage, setErrorMessage] = useState('');
-
-    const schema = yup.object({
-        boardSize: yup.number()
-            .required('This field is required.')
-            .positive('Number must be positive.')
-            .integer('Number must be integer.')
-            .min(3, 'Min value 3.')
-            .max(7, 'Max value  7.')
-            .nullable()
-    });
-
-    schema
-        .validate({boardSize})
-        .then(function(value) {
-            // console.log(value);
-        })
-        .catch(function(err) {
-            setError(true);
-            setErrorMessage(err.message);
-            // console.log('error message = ', err.message);
-        });
-
-
     return (
         <>
-            {error && <h3>{errorMessage}</h3>}
-
             <form onSubmit={handleChangeBoardSize}>
                 <label>Board size</label>
                 <input
                     type="number"
+                    min={BOARD_SIZE}
+                    max={10}
                     value={boardSize}
                     onChange={e => setBoardSize(e.target.value)}
                 />
 
-                {!winner && <button>Change</button>}
+                {!winner && <button>Change board size</button>}
             </form>
         </>
     );
